Server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
exceptiongroup==1.1.3
expertai-nlapi==2.5.0
fastapi==0.103.1
firebase-admin==6.2.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
frozenlist==1.4.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-language==2.11.0
google-cloud-storage==2.10.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-generativeai==0.1.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
nylas==5.14.1
oauthlib==3.2.2
openai==0.28.0
packaging==23.2
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.16
URLObject==2.4.3
uvicorn==0.23.2
websocket-client==0.59.0
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
